[FIX] compilator warning
warning was made because I forogot to cast lineThick in raylibCpp.cpp into function "DrawRectangleLinesEx" (line 15)

+++ src/raylibCpp.cpp
@@ -19,5 +19,5 @@ void raycpp::DrawRectangleLinesEx(Vec2<int> pos, Vec2<int> widthHeight, float li
     assert(lineThick > 0);
     DrawRectangleLinesEx({(float)pos.GetX(), (float)pos.GetY(),
                           (float)widthHeight.GetX(), (float)widthHeight.GetY()},
-                         lineThick, color);
+                         (float)lineThick, color);
 }
